Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
langchain
langchain-openai
langchain-chroma
langchain-community
langchain-huggingface
openai
chromadb
sentence-transformers
